+++ tasks/seq/laganina_e_sum_values_by_columns_matrix/func_tests/main.cpp
@@ -1,9 +1,19 @@
 
 #include <gtest/gtest.h>
-
+#include <random>
 #include <vector>
 
 #include "seq/laganina_e_sum_values_by_columns_matrix/include/ops_seq.hpp"
+
+std::vector<int> laganina_e_sum_values_by_columns_matrix_seq::getRandomVector(int sz) {
+  std::random_device dev;
+  std::mt19937 gen(dev());
+  std::vector<int> vec(sz);
+  for (int i = 0; i < sz; i++) {
+    vec[i] = (gen() % 100) - 49;
+  }
+  return vec;
+}
 
 TEST(laganina_e_sum_values_by_columns_matrix_seq, Test_2_2_matrix) {
   int n = 2;
